fix(hero.js): addressing removal icons, pass image and title as children
Removed the icons. The image and title of the hero widget are now passed as children to the
component.
